refactor(DPGO-19): add appropriate pasw usecase
